feat: refactor histplot + new get_plottables function
